improvement(doc): reactlive toolbar extraction and codesandboxbutton moving
