+++ packages/doc/src/components/CodeBlock/ReactLive/CodeSandboxButton.jsx
@@ -1,16 +1,18 @@
-import React, { useState } from 'react';
-import { string, arrayOf } from 'prop-types';
+import React, { useState, useContext } from 'react';
 import CodeSandboxIcon from 'images/codesandbox.svg';
 import LoadingIcon from 'images/loading.svg';
+import ReactLiveContext from './ReactLiveContext';
+import { URL, setOptions } from './sandboxConfig';
 
-import { URL, setOptions } from './code';
-
-const CodeSandboxButton = ({ code }) => {
+const CodeSandboxButton = () => {
   const [loading, setLoading] = useState(false);
+  const { imports, code } = useContext(ReactLiveContext);
+
+  const sandbox = [imports, code];
 
   const createSandbox = () => {
     setLoading(true);
-    fetch(URL, setOptions(code))
+    fetch(URL, setOptions(sandbox))
       .then(x => x.json())
       .then(data =>
         window.open(
@@ -30,10 +32,6 @@ const CodeSandboxButton = ({ code }) => {
       )}
     </>
   );
-};
-
-CodeSandboxButton.propTypes = {
-  code: arrayOf(string).isRequired,
 };
 
 export default CodeSandboxButton;
