+++ packages/doc/src/components/CodeBlock/ReactLive/ReactLiveContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const ReactLiveContext = createContext({});
+
+export default ReactLiveContext;
