+++ packages/doc/src/components/CodeBlock/ReactLive/sandboxConfig.js
