--- packages/doc/src/components/CodeSandboxButton/index.js
@@ -1,3 +1,0 @@
-import CodeSandboxButton from './CodeSandboxButton';
-
-export default CodeSandboxButton;
