+++ packages/doc/src/components/index.js
@@ -12,7 +12,6 @@ import Documentation from './Documentation';
 import Dropdown from './Dropdown';
 import { TabbedView, Tab } from './TabbedView';
 import TabBar from './TabBar';
-import CodeSandboxButton from './CodeSandboxButton';
 import ThemeConfig from './ThemeConfig';
 import GlobalStyle from './GlobalStyle';
 import Redirect from './Redirect';
@@ -38,7 +37,6 @@ export {
   TabbedView,
   Tab,
   TabBar,
-  CodeSandboxButton,
   ThemeConfig,
   GlobalStyle,
   Redirect,
